app/requirements.txt
flask
grpcio-tools
mysql-connector
keras
numpy
h5py
image
pandas